# Reduce `Scheduler` overhead when no tasks are pending