--- a/osu.Framework/Threading/Scheduler.cs
+++ b/osu.Framework/Threading/Scheduler.cs
@@ -123,7 +123,6 @@ namespace osu.Framework.Threading
 
             while (getNextTask(out ScheduledDelegate sd))
             {
-                //todo: error handling
                 sd.RunTaskInternal();
 
                 TotalTasksRun++;