@peppy 2022-07-07
> Interestingly, I don't see the performance regression mentioned in OP upon running the benchmark on my machine:

I can still repro the difference... will try and figure it out since it's a bit alarming to me.