@smoogipoo 2022-07-07
I can also repro a little bit of slowdown (300us):

Master:

|                         Method |        Mean |      Error |    StdDev |  Gen 0 | Allocated |
|------------------------------- |------------:|-----------:|----------:|-------:|----------:|
|           UpdateEmptyScheduler |    65.01 us |   0.171 us |  0.009 us |      - |         - |
| UpdateSchedulerWithManyDelayed | 1,452.63 us | 263.561 us | 14.447 us |      - |       3 B |
| UpdateSchedulerWithEveryUpdate |   113.26 us |   1.446 us |  0.079 us |      - |         - |
|   UpdateSchedulerWithManyAdded |   114.33 us |  23.513 us |  1.289 us | 0.8545 |  80,000 B |

This PR:
|                         Method |         Mean |       Error |     StdDev |  Gen 0 | Allocated |
|------------------------------- |-------------:|------------:|-----------:|-------:|----------:|
|           UpdateEmptyScheduler |     2.893 us |   0.1699 us |  0.0093 us |      - |         - |
| UpdateSchedulerWithManyDelayed | 1,707.047 us | 232.6123 us | 12.7503 us |      - |       3 B |
| UpdateSchedulerWithEveryUpdate |   115.545 us |  10.8061 us |  0.5923 us |      - |         - |
|   UpdateSchedulerWithManyAdded |    82.162 us |   3.2974 us |  0.1807 us | 0.8545 |  80,000 B |

---

Seems to be caused by the removal of the count check in `updateTimedTasks`.