@peppy 2022-07-07
How can that be the case though? I explicitly added the count check (removed in the last commit) to ensure it wasn't contributing to anything. Can you test with that commit reverted and comment on why the location being different matters if you have an idea?