@smoogipoo 2022-07-07
Reverting the last commit doesn't fix it. The only thing that seems to fix it is moving the check inside.

I have a feeling this is more about branch prediction hinting. I did some testing on this a while back to see how C# behaves in the context of early returns, but I'm trying to remember the specifics...